+++ apps/frontend/src/features/admin-form/settings/SettingsWebhooksPage.tsx
@@ -1,36 +1,20 @@
-import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useParams } from 'react-router-dom'
 import { Skeleton } from '@chakra-ui/react'
-import { useFeatureIsOn, useGrowthBook } from '@growthbook/growthbook-react'
+import { useFeatureIsOn } from '@growthbook/growthbook-react'
 
 import { featureFlags } from 'formsg-shared/constants'
 import { FormResponseMode } from 'formsg-shared/types/form'
-import { FormId } from 'formsg-shared/types/form/form'
-
-import { useUser } from '~features/user/queries'
 
 import { CategoryHeader } from './components/CategoryHeader'
 import { WebhooksSection } from './components/WebhooksSection'
 import { WebhooksUnsupportedMsg } from './components/WebhooksSection/WebhooksUnsupportedMsg'
-import { WebhookV1SchemaInfobox } from './components/WebhooksSection/WebhookV1SchemaInfobox'
 import { useAdminFormSettings } from './queries'
 
 export const SettingsWebhooksPage = (): JSX.Element => {
   const { t } = useTranslation()
-  const gb = useGrowthBook()
   const { data: settings, isLoading } = useAdminFormSettings()
-  const userRes = useUser()
 
-  useEffect(() => {
-    gb?.setAttributes({
-      adminEmail: userRes.user?.email,
-    })
-  }, [userRes.user?.email, gb])
-
-  const { formId } = useParams()
   const enableMrfWebhooks = useFeatureIsOn(featureFlags.enableMrfWebhooks)
-  const isMrfCutoverEnabled = useFeatureIsOn(featureFlags.mrfCutover)
 
   const enableWebhooks =
     !isLoading &&
@@ -47,19 +31,11 @@ export const SettingsWebhooksPage = (): JSX.Element => {
     )
   }
 
-  const showV1SchemaInfobox =
-    isMrfCutoverEnabled &&
-    settings?.responseMode === FormResponseMode.Encrypt &&
-    !!formId
-
   return (
     <Skeleton isLoaded={!isLoading}>
       <CategoryHeader>
         {t('features.adminForm.settings.webhooks.title')}
       </CategoryHeader>
-      {showV1SchemaInfobox && (
-        <WebhookV1SchemaInfobox formId={formId as FormId} />
-      )}
       <WebhooksSection />
     </Skeleton>
   )