+++ apps/frontend/src/features/workspace/components/CreateFormModal/CreateFormModalContent/CreateFormStorageModeScreen.tsx
@@ -30,16 +30,14 @@ export const CreateFormStorageModeScreen = (): JSX.Element => {
               onClick={goToMrfDetails}
               mr="0.5rem"
             />
-            Set up a Storage mode form
+            {t('features.workspace.modals.forms.create.legacy.title')}
           </Flex>
         </Container>
       </ModalHeader>
       <ModalBody whiteSpace="pre-wrap">
         <Container maxW="45rem" p={0}>
           <Text textStyle="body-2" color="secondary.500" mb="2rem">
-            Storage mode is outdated and no longer receives new features. Only
-            use this if you need a feature not yet available in the current
-            version.
+            {t('features.workspace.modals.forms.create.legacy.description')}
           </Text>
           <FormTitleInput mb="2.5rem" />
           <Button