+++ apps/frontend/src/features/workspace/components/WorkspaceSearchbar/WorkspaceSearchbar.tsx
@@ -20,7 +20,10 @@ import Button from '~components/Button'
 import Menu from '~components/Menu'
 
 import { FilterOption } from '~features/workspace/types'
-import { FILTER_OPTIONS } from '~features/workspace/utils/dashboardFilter'
+import {
+  getFilterOptionLabel,
+  useDashboardFilterOptions,
+} from '~features/workspace/utils/dashboardFilter'
 
 export interface WorkspaceSearchbarProps {
   /** Filter selection input will be controlled if provided. */
@@ -106,6 +109,8 @@ export const WorkspaceSearchbar = forwardRef<WorkspaceSearchbarProps, 'input'>(
       onFilter,
     })
 
+    const filterOptions = useDashboardFilterOptions()
+
     const filterRef = useRef<HTMLDivElement>(null)
 
     const [filterElemWidth, setFilterElemWidth] = useState<number>()
@@ -160,9 +165,9 @@ export const WorkspaceSearchbar = forwardRef<WorkspaceSearchbarProps, 'input'>(
                   value={internalFilter}
                   onChange={(val) => setInternalFilter(val as FilterOption)}
                 >
-                  {FILTER_OPTIONS.map((value, i) => (
+                  {filterOptions.map((value, i) => (
                     <MenuItemOption key={i} iconSpacing="1.5rem" value={value}>
-                      {value}
+                      {getFilterOptionLabel(value, t)}
                     </MenuItemOption>
                   ))}
                 </MenuOptionGroup>