+++ apps/frontend/src/i18n/locales/features/admin-form/sidebar/en-sg.ts
@@ -1,6 +1,7 @@
 import { enSG as fields } from './fields'
 import { enSG as headerAndInstructions } from './header-and-instructions'
 import { enSG as logic } from './logic'
+import { enSG as navLabels } from './nav-labels'
 import { enSG as thankYou } from './thank-you'
 import { enSG as workflow } from './workflow'
 
@@ -8,6 +9,7 @@ export const enSG = {
   fields,
   headerAndInstructions,
   logic,
+  navLabels,
   thankYou,
   workflow,
 }